src/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
neo4j==5.19.0
faiss-cpu==1.13.0
sentence-transformers==2.6.1
pydantic
pydantic-settings
